Added optional "align" param to Malloc helpers (default 16 bytes)

+++ VulkanGraphicsAPI/Typedefs.hpp
@@ -2,7 +2,7 @@
 
 namespace Utility
 {
-    typedef uint32_t      UIndex_t;
-    typedef int32_t       Index_t; 
+    typedef uint8_t       UIndex_t;
+    typedef int8_t        Index_t; 
     typedef unsigned char NeighbourCount_t;
 }
